mediaserver: Split segment handling into processSegment.
